Finished the routes page
This should when a user has a route saved to the local database, upon tapping the route tab diaplay the route name and any pertinant info about the route inside a card view inside a recycler view. as of right now nothing currently disaplays. Because there are no routes on the local database and to add one my versions location set button would have to work but it does not.